LayoutTests/imported/w3c/web-platform-tests/css/cssom/mediaquery-sort-dedup-expected.txt


FAIL Media queries are not sorted lexicographically assert_equals: expected "(min-width: 10px) and (min-height: 10px)" but got "(min-height: 10px) and (min-width: 10px)"
FAIL Media queries are not deduplicated assert_equals: expected "(color) and (color)" but got "(color)"
